feat: add deleteTask function in taskController.js

+++ src/controllers/taskController.js
@@ -1,9 +1,5 @@
-import { Sequelize } from "sequelize";
 import { Board } from "../models/Board.js";
 import { Task } from "../models/Task.js";
-import { User } from "../models/User.js";
-import { sequelize } from "../db/db.js";
-import { Board_Task } from "../models/Board_Task.js";
 
 
 class TaskController {
@@ -36,28 +32,7 @@ class TaskController {
     async getTasks(req, res) {
         try {
 
-            /* const boards = board.map(item => {
-                return {
-                    boardId: item.boardId,
-                    title: item.title,
-
-
-                    items: tasks.map(async (child) => {
-                        if (child.board_id === item.id) {
-                            console.log({ task_id: child.task_id, content: child.content, author: await getAuthor(child.user_id) })
-                            await Promise.all(arr.map(async el => ++el))
-                            return {
-                                task_id: child.task_id, content: child.content, author: await getAuthor(child.user_id)
-                            }
-
-                        }
-                    }).filter(child => child !== null && child !== undefined)
-                }
-            }); */
-
-
             const boards = await Board.findAll({ include: Task });
-            console.log(JSON.stringify(boards, null, 2))
 
             res.json(boards)
         } catch (error) {
@@ -73,9 +48,12 @@ class TaskController {
         try {
             const post_id = +req.params.post_id;
 
+            if (!post_id) {
+                throw new Error("Некорректный id")
+            }
 
             const deletePost = await Task.destroy({ where: { post_id } })
-            res.json("успешно удален")
+            res.json({ message: "Успешно удален", target: deletePost })
         } catch (error) {
             res.status(400).json({ message: error.message, error: error })
         }
